+++ src/main/java/backend/zip/service/brokeritem/BrokerItemOptionService.java
@@ -1,4 +1,4 @@
-package backend.zip.service.broker;
+package backend.zip.service.brokeritem;
 
 import backend.zip.domain.broker.BrokerOption;
 import backend.zip.dto.brokeritem.request.AddBrokerItemOptionsRequest;
