+++ src/main/java/backend/zip/service/brokeritem/BrokerItemService.java
@@ -1,4 +1,4 @@
-package backend.zip.service.broker;
+package backend.zip.service.brokeritem;
 
 import backend.zip.domain.broker.BrokerItem;
 import backend.zip.dto.brokeritem.request.AddBrokerItemDetailsRequest;
@@ -6,8 +6,6 @@ import backend.zip.dto.brokeritem.request.AddBrokerItemOptionsRequest;
 import backend.zip.dto.brokeritem.response.BrokerItemAddressResponse;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 public interface BrokerItemService {
     BrokerItem saveBrokerItem(Long userId, BrokerItem brokerItem, BrokerItemAddressResponse brokerItemAddressResponse,
